pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fieldsense-research"
version = "1.0.0"
description = "Flask UI and API for crop, fertilizer, and leaf agricultural ML models."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "flask>=3.0.0",
  "numpy>=1.24.0,<2",
  "scikit-learn>=1.3.0",
  "category-encoders>=2.6.0",
  "Pillow>=10.0.0",
  "pandas>=2.0.0",
  "onnxruntime>=1.16.0,<2"
]

[tool.setuptools.packages.find]
where = ["src"]
